Move single-cluster specific fixtures to Kuadrant/conftest.py
